Removed AbstractService, renamed ServiceInterface to CrudServiceInterface
